fix: forward vite plugin config to esbuild plugin
